Initial creation
I was interrupted when I was making this so I haven't run any of the code; I wanted to get it pushed to Github all the same. It's probably a good time to do so anyway since this should be the basic outline for all three files.

index.html
- Setup basic outline for code with a container for the etch a sketch

style.css
- Setup the CSS for the container and the boxes (for now hardcoded at 60px)

index.js
- Basic code for creating a grid of flex boxes to make the grid
 - Odin isn't clear whether this should be one flex row which wraps or if it can be multiple ones
 - Since the boxes don't relate to each other it seems easier to just make a bunch of "flex rows" instead of one big flex box